requirements.txt
streamlit
python-dotenv
langchain-anthropic
langchain-core
langchain-community
langchain-text-splitters
pypdf